+++ api/partimark_app/models/system_config.py
@@ -22,7 +22,7 @@ class SystemConfig(Base):
         index=True,
     )
 
-    max_weekly_score: Mapped[float] = mapped_column(Float, nullable=False, default=3.0)
+    max_weekly_score: Mapped[int] = mapped_column(Integer, nullable=False, default=3.0)
     total_participation_points: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
 
     is_configured: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
